[xbmc] move CEnvironment from utils to platform
