+++ xbmc/platform/Environment.cpp
