+++ xbmc/platform/Environment.h
